BUG#26080 - Memory Storage engine not working properly
Extending varchar column length with ALTER TABLE may result in unusable
memory table.

The problem is that we use fast ALTER TABLE in this case, which is not
supported by now.

This is fixed by refusing fast ALTER TABLE when extending varchar column.
In other words force copy of a table during ALTER TABLE.

Affects MEMORY tables in 5.1 only.