Bug #28366: multiple load_file('a: ... ') causes entire mysqld process to hang
When a Windows console application that has an open console (e.g. mysqld-nt 
started with the --console option) encounters certain type of errors 
(like no floppy disk in a floppy drive) the OS will pop-up an 
"abort/retry/ignore" dialog and block the application (depending on a 
registry setting : see http://msdn2.microsoft.com/en-us/embedded/aa731206.aspx
for details).
Fixed by disabling the dialog popups for every error except a GPF and
alignment errors. This is safe to do as the actual error gets reported
(and handled) to mysqld.