BUG#30274 (Debug server crashes because mutex isn't initialized):
Completion patch, since it is possible that Log_event::write_header()
is called with no current thread.

When starting the server, there is no thread available, but
MYSQL_BIN_LOG::open() is called to initialize the binary log. The
function writes a format description log event to the binary log,
which means that Log_event::write_header() can be called with no
current thread object available.