Bug #30393: Test "group_by" fails with a difference in "row count"     and strategy (explain)
The fix for WL3527 adds tests that test if the index usage hints
combinations don't cause syntax errors.
The EXPLAIN for one of these tests can be affected by the size of the
rowid on the disk (affected by the presence of large file support). 
Fixed to avoid the platform dependent test result by removing the 
irrelevant columns from the EXPLAIN result.